+++ Spring/src/main/java/bitc/full502/final_project_team1/core/service/UserStatusService.java
@@ -3,5 +3,5 @@ package bitc.full502.final_project_team1.core.service;
 import bitc.full502.final_project_team1.api.app.dto.AppUserSurveyStatusResponse;
 
 public interface UserStatusService {
-    AppUserSurveyStatusResponse getUserStatus(Integer userId);
+    AppUserSurveyStatusResponse getUserStatus(Long userId);
 }
